+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 Yes, the irony is not lost on me. :)
 
+## 2.0.1
+
+- Fix [#33](https://github.com/planttheidea/unchanged/issues/33) - ensure objects created with `Object.create(null)` do not error
+- Switch from `babel-minify` to `terser` for minified build
+
 ## 2.0.0
 
 Rewrite in TypeScript!